+++ Tests/Controller/ControllerTrait.php
@@ -58,12 +58,12 @@ trait ControllerTrait
         $container->setParameter('dtc_queue.class.run', $runManager->getRunClass());
         $container->setParameter('dtc_queue.class.run_archive', $runManager->getRunArchiveClass());
         $container->setParameter('dtc_queue.admin.chartjs', '');
-        $container->setParameter('dtc_queue.default_manager', 'orm');
-        $container->setParameter('dtc_queue.record_timings', true);
-        $container->setParameter('dtc_queue.record_timings_timezone_offset', 0);
-        $container->set('dtc_queue.job_manager', $jobManager);
-        $container->set('dtc_queue.job_timing_manager', $jobTimingManager);
-        $container->set('dtc_queue.worker_manager', new WorkerManager($jobManager, new EventDispatcher()));
+        $container->setParameter('dtc_queue.manager.job', 'orm');
+        $container->setParameter('dtc_queue.timings.record', true);
+        $container->setParameter('dtc_queue.timings.timezone_offset', 0);
+        $container->set('dtc_queue.manager.job', $jobManager);
+        $container->set('dtc_queue.manager.job_timing', $jobTimingManager);
+        $container->set('dtc_queue.manager.worker', new WorkerManager($jobManager, new EventDispatcher()));
         $rendererFactory = new RendererFactory(
             new Router(new YamlFileLoader(new FileLocator(__DIR__)), 'test.yml'),
             [
@@ -90,7 +90,7 @@ trait ControllerTrait
         $liveJobsGridSource = new $liveJobsGridSourceClass($jobManager);
         $liveJobsGridSource->setRunning(true);
         $container->set('dtc_queue.grid_source.jobs_running.orm', $liveJobsGridSource);
-        $container->set('dtc_queue.job_manager', $jobManager);
+        $container->set('dtc_queue.manager.job', $jobManager);
         $gridSourceManager = new GridSourceManager(new AnnotationReader(), __DIR__);
         $container->set('dtc_grid.manager.source', $gridSourceManager);
         $gridSourceJob = new $gridSourceClass($jobManager->getObjectManager(), $jobManager->getJobClass());